requirements.txt
selenium
RealtimeSTT
pygame
google
google-genai
pydantic
pillow 
pywin32
openai
boto3
botocore
sounddevice
websocket-client==1.8.0
websockets
pyaudio
python-dotenv
chromadb 
sentence-transformers
groq